# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Market API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Car {
  id: string;
  brand: string;
  model: string;
  year: number;
  price: number;
  mileage: number;
  color: string;
  body_type: string;
  condition: string;
  status: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  features?: string[] | null;
  /** @nullable */
  images?: string[] | null;
  /** @nullable */
  added_by?: string | null;
  created_at: string;
  updated_at: string;
}

export interface CarInput {
  brand: string;
  model: string;
  year: number;
  price: number;
  mileage: number;
  color: string;
  body_type: string;
  condition?: string;
  status?: string;
  description?: string;
  features?: string[];
  images?: string[];
}

export interface CarUpdate {
  brand?: string;
  model?: string;
  year?: number;
  price?: number;
  mileage?: number;
  color?: string;
  body_type?: string;
  condition?: string;
  status?: string;
  description?: string;
  features?: string[];
  images?: string[];
}

export interface CarStats {
  total: number;
  for_sale: number;
  sold: number;
  brands_count: number;
  avg_price: number;
}

export interface Profile {
  id: string;
  user_id: string;
  /** @nullable */
  display_name?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  avatar_url?: string | null;
  created_at: string;
  updated_at: string;
}

export interface ProfileUpdate {
  display_name?: string;
  phone?: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  display_name?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthUser {
  id: string;
  email: string;
  /** @nullable */
  display_name?: string | null;
}

export interface AuthResponse {
  user: AuthUser;
  token: string;
}

export type GetCarsParams = {
  brand?: string;
  body_type?: string;
  condition?: string;
  status?: string;
  min_price?: number;
  max_price?: number;
  min_year?: number;
  max_year?: number;
};