# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Car Market API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: cars
    description: Car listings
  - name: favorites
    description: User favorites
  - name: profiles
    description: User profiles
  - name: auth
    description: Authentication
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /cars:
    get:
      operationId: getCars
      tags: [cars]
      summary: List all cars
      parameters:
        - name: brand
          in: query
          schema: { type: string }
        - name: body_type
          in: query
          schema: { type: string }
        - name: condition
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: min_price
          in: query
          schema: { type: integer }
        - name: max_price
          in: query
          schema: { type: integer }
        - name: min_year
          in: query
          schema: { type: integer }
        - name: max_year
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of cars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"
    post:
      operationId: createCar
      tags: [cars]
      summary: Create a new car listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
        "401":
          description: Unauthorized

  /cars/stats:
    get:
      operationId: getCarStats
      tags: [cars]
      summary: Get car marketplace statistics
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarStats"

  /cars/{id}:
    get:
      operationId: getCar
      tags: [cars]
      summary: Get a car by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Car details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
        "404":
          description: Not found
    patch:
      operationId: updateCar
      tags: [cars]
      summary: Update a car listing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteCar
      tags: [cars]
      summary: Delete a car listing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /favorites:
    get:
      operationId: getFavorites
      tags: [favorites]
      summary: Get current user favorites
      responses:
        "200":
          description: List of favorite car IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "401":
          description: Unauthorized

  /favorites/{carId}:
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add car to favorites
      parameters:
        - name: carId
          in: path
          required: true
          schema: { type: string }
      responses:
        "201":
          description: Added
        "401":
          description: Unauthorized
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove car from favorites
      parameters:
        - name: carId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removed
        "401":
          description: Unauthorized

  /profile:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get current user profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateProfile
      tags: [profiles]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "204":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated


components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Car:
      type: object
      required: [id, brand, model, year, price, mileage, color, body_type, condition, status, created_at, updated_at]
      properties:
        id:         { type: string }
        brand:      { type: string }
        model:      { type: string }
        year:       { type: integer }
        price:      { type: integer }
        mileage:    { type: integer }
        color:      { type: string }
        body_type:  { type: string }
        condition:  { type: string }
        status:     { type: string }
        description: { type: ["string", "null"] }
        features:   { type: ["array", "null"], items: { type: string } }
        images:     { type: ["array", "null"], items: { type: string } }
        added_by:   { type: ["string", "null"] }
        created_at: { type: string }
        updated_at: { type: string }

    CarInput:
      type: object
      required: [brand, model, year, price, mileage, color, body_type]
      properties:
        brand:       { type: string }
        model:       { type: string }
        year:        { type: integer }
        price:       { type: integer }
        mileage:     { type: integer }
        color:       { type: string }
        body_type:   { type: string }
        condition:   { type: string }
        status:      { type: string }
        description: { type: string }
        features:    { type: array, items: { type: string } }
        images:      { type: array, items: { type: string } }

    CarUpdate:
      type: object
      properties:
        brand:       { type: string }
        model:       { type: string }
        year:        { type: integer }
        price:       { type: integer }
        mileage:     { type: integer }
        color:       { type: string }
        body_type:   { type: string }
        condition:   { type: string }
        status:      { type: string }
        description: { type: string }
        features:    { type: array, items: { type: string } }
        images:      { type: array, items: { type: string } }

    CarStats:
      type: object
      required: [total, for_sale, sold, brands_count, avg_price]
      properties:
        total:        { type: integer }
        for_sale:     { type: integer }
        sold:         { type: integer }
        brands_count: { type: integer }
        avg_price:    { type: integer }

    Profile:
      type: object
      required: [id, user_id, created_at, updated_at]
      properties:
        id:           { type: string }
        user_id:      { type: string }
        display_name: { type: ["string", "null"] }
        phone:        { type: ["string", "null"] }
        avatar_url:   { type: ["string", "null"] }
        created_at:   { type: string }
        updated_at:   { type: string }

    ProfileUpdate:
      type: object
      properties:
        display_name: { type: string }
        phone:        { type: string }

    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email:        { type: string }
        password:     { type: string }
        display_name: { type: string }

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:    { type: string }
        password: { type: string }

    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:  { $ref: "#/components/schemas/AuthUser" }
        token: { type: string }

    AuthUser:
      type: object
      required: [id, email]
      properties:
        id:           { type: string }
        email:        { type: string }
        display_name: { type: ["string", "null"] }