# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Market API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from "./authUser";

export interface AuthResponse {
  user: AuthUser;
  token: string;
}