# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Market API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: string;
  email: string;
  /** @nullable */
  display_name?: string | null;
}