# lib/api-zod/src/generated/types/car.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Market API
 * OpenAPI spec version: 0.1.0
 */

export interface Car {
  id: string;
  brand: string;
  model: string;
  year: number;
  price: number;
  mileage: number;
  color: string;
  body_type: string;
  condition: string;
  status: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  features?: string[] | null;
  /** @nullable */
  images?: string[] | null;
  /** @nullable */
  added_by?: string | null;
  created_at: string;
  updated_at: string;
}