# lib/api-zod/src/generated/types/carInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Market API
 * OpenAPI spec version: 0.1.0
 */

export interface CarInput {
  brand: string;
  model: string;
  year: number;
  price: number;
  mileage: number;
  color: string;
  body_type: string;
  condition?: string;
  status?: string;
  description?: string;
  features?: string[];
  images?: string[];
}