# lib/api-zod/src/generated/types/carStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Market API
 * OpenAPI spec version: 0.1.0
 */

export interface CarStats {
  total: number;
  for_sale: number;
  sold: number;
  brands_count: number;
  avg_price: number;
}