# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Market API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: string;
  user_id: string;
  /** @nullable */
  display_name?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  avatar_url?: string | null;
  created_at: string;
  updated_at: string;
}